First attempt of implementing an HTML-parser which is capable to compile HTML input into DOM-objects of the html5 library, and an extra-feature to bind them to their root node for further access.
This attempt makes it possible to create PyJS apps using the HTML5 library without creating every single element by hand.

A first attempt of this parser was written using re, unfortunatelly the re from PyJS are not complying to the re's of Python standard, so the feature was removed and all is done by hand now.